Extract bootloader config into a separate header
