+++ src/libaktualizr/bootloader/bootloader.h
@@ -1,17 +1,7 @@
 #ifndef BOOTLOADER_H_
 #define BOOTLOADER_H_
 
-#include <boost/property_tree/ini_parser.hpp>
-#include <iostream>
-
-enum RollbackMode { kBootloaderNone = 0, kUbootGeneric, kUbootMasked };
-
-struct BootloaderConfig {
-  RollbackMode rollback_mode{kBootloaderNone};
-
-  void updateFromPropertyTree(const boost::property_tree::ptree& pt);
-  void writeToStream(std::ostream& out_stream) const;
-};
+#include "bootloader_config.h"
 
 class Bootloader {
  public:
